requirements.txt
streamlit==1.40.1
requests==2.32.3
Pillow==11.0.0